refactor interaction files
